// src/checkout/interfaces/checkout-types.ts
export enum CheckoutBlockerCode {
  ADDRESS_NOT_SERVICEABLE = 'ADDRESS_NOT_SERVICEABLE',
  CART_EMPTY = 'CART_EMPTY',
  CART_RESTAURANT_MISMATCH = 'CART_RESTAURANT_MISMATCH',
  ITEM_INACTIVE = 'ITEM_INACTIVE',
  ITEM_UNAVAILABLE = 'ITEM_UNAVAILABLE',
  INVALID_ITEM_QUANTITY = 'INVALID_ITEM_QUANTITY',
  MINIMUM_ORDER_NOT_MET = 'MINIMUM_ORDER_NOT_MET',
  PRICE_CHANGED = 'PRICE_CHANGED',
  RESTAURANT_CLOSED = 'RESTAURANT_CLOSED',
  RESTAURANT_UNAVAILABLE = 'RESTAURANT_UNAVAILABLE',
}

export interface CheckoutBlocker {
  code: CheckoutBlockerCode;
  message: string;
}